✨ Establishing the domain model (#12)
Added all main data models that we plan to expose for managing for our users.
Implemented the matcher to find friend/location references in a free-form text in the new activity command.
Renamed Life to Journal to make it more standard.
Added a way to parse human-readable dates like yesterday, a day before, etc.